+++ DBUtils/Dao/DaoRecordsetImpl.h
@@ -48,8 +48,8 @@ public:
 
 	virtual bool IsEOF() override;
 
-	virtual void SetFieldBinary(const wchar_t *sFieldName, unsigned char *pData, unsigned long nSize) override;
-	virtual void GetFieldBinary(const wchar_t *sFieldName, unsigned char **pData, unsigned long &nSize) override;
+	virtual void SetFieldBinary(const wchar_t *sFieldName, unsigned char *pData, size_t nSize) override;
+	virtual void GetFieldBinary(const wchar_t *sFieldName, unsigned char **pData, size_t &nSize) override;
 	virtual void FreeBinary(unsigned char *pData) override;
 
     virtual void SetFieldValueNull(const wchar_t *lpszName) override;
